+++ server/src/interfaces/Notification.ts
@@ -1,0 +1,9 @@
+interface Notification {
+  id: string;
+  title: string;
+  text: string;
+  href: string;
+  user_id: string;
+}
+
+export default Notification;