+++ server/src/lib/database.ts
@@ -34,6 +34,28 @@ async function select1() {
 
 async function updateDb() {
   try {
+    await processQuery(`
+    ALTER TABLE \`911calls\` CHANGE \`assigned_unit\` \`assigned_unit\` TEXT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL; 
+    --
+    -- Table structure for table \`notifications\`
+    --
+    
+    CREATE TABLE \`notifications\` (
+      \`id\` varchar(255) NOT NULL,
+      \`title\` varchar(255) NOT NULL,
+      \`text\` text NOT NULL,
+      \`href\` text NOT NULL,
+      \`user_id\` varchar(255) NOT NULL
+    ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+    
+    -- --------------------------------------------------------
+    
+    --
+    -- Indexes for table \`notifications\`
+    --
+    ALTER TABLE \`notifications\`
+      ADD PRIMARY KEY (\`id\`);
+    `).catch();
     await processQuery(
       `
       CREATE TABLE \`court_requests\` (
@@ -52,8 +74,8 @@ async function updateDb() {
       --
       ALTER TABLE \`court_requests\`
         ADD PRIMARY KEY (\`id\`);
-      ALTER TABLE \`citizens\` ADD \`note\` VARCHAR(255) NOT NULL AFTER \`b_status\`;ALTER TABLE \`officers\` ADD \`callsign\` VARCHAR(255) NOT NULL AFTER \`officer_dept\`;`
-    );
+      ALTER TABLE \`citizens\` ADD \`note\` VARCHAR(255) NOT NULL AFTER \`b_status\`;ALTER TABLE \`officers\` ADD \`callsign\` VARCHAR(255) NOT NULL AFTER \`officer_dept\`;`,
+    ).catch();
 
     // eslint-disable-next-line no-empty
   } catch {}